common.utils: Provide hooks
Bot specific code moved to `lrrbot.decorators`.
